# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TAO Test Center API",
    "version": "0.1"
  },
  "paths": {
...
    "/taoTestCenter/api/testCenter": {
      "get": {
        "tags": [
          "testCenter"
        ],
        "summary": "Get test center data",
        "description": "Get test center data",
        "operationId": "oat\\taoTestCenter\\controller\\RestTestCenter::get",
        "parameters": [
          {
            "name": "testCenter",
            "in": "query",
            "description": "testCenter Uri (Url encoded)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test center data",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "label": {
                      "description": "Test center label",
                      "type": "string"
                    },
                    "class": {
                      "description": "Test center class URI",
                      "type": "string"
                    }
                  },
                  "type": "object",
                  "example": {
                    "label": "Test Center ABC",
                    "class": "http://sample/first.rdf#i15367360596713165"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid test center Uri",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "errorCode": 404,
                    "errorMsg": "Test Center `http://sample/first.rdf#i15367360596713165` does not exist.",
                    "version": "3.3.0-sprint85"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "testCenter"
        ],
        "summary": "Create new test center",
        "description": "Create new test center",
        "operationId": "oat\\taoTestCenter\\controller\\RestTestCenter::post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "label"
                ],
                "properties": {
                  "class-uri": {
                    "description": "Class uri to import item. If not specified root class will be used.",
                    "type": "string"
                  },
                  "class-label": {
                    "description": "Label of class to import item. If not specified root class will be used.\n   * If label is not unique first match will be used.",
                    "type": "string"
                  },
                  "label": {
                    "description": "Test center label",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created test center URI",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "description": "`false` on failure, `true` on success",
                      "type": "boolean"
                    },
                    "uri": {
                      "description": "Created test center URI",
                      "type": "string"
                    }
                  },
                  "type": "object",
                  "example": {
                    "success": true,
                    "uri": "http://sample/first.rdf#i1536680377163171"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid class uri",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "errorCode": 0,
                    "errorMsg": "Class does not exist. Please use valid class-uri or class-label",
                    "version": "3.3.0-sprint85"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}